requirements.txt
discord.py==2.1.0
openai
python-dotenv==0.21.0
requests==2.25.1
langchain
